Further restructure
Make it clear that as we recurse through the AbcTune, building our state
called TState, we don't use the final value from the state monad, we
only use the returned TState (from execState).

The state holds a list of MidiBar which contains the flat desctiption of
the tune together with information about repeated sections.  We return
this state using execState and then generate the Melody with all repeats
instantiated, properly phrased.