Replace isNullSection with isDeadSection..
in the RepeatSection module.  This now labels a section as effectively
null if there is no repeat and if the start position is greater than or
equal to the end position (taking account of nulls).  This is more
accurate and should ensure that when we gather together all the
sections, they are all valid ones.