Generate some sort of intro
The infrastructure is now largely in place.  The PlayableAbc class now
has the concept of an optional intro.  An intro is indeed generated but
at the moment it's incorrect.

The main problem is that we have two references to the bars we're using
as an intro.  One for the main part of the tune, the other for their reuse as an intro in a completely new section.  But of course the note offsets are wrong because we have to reset to zero whenever we start a new section.

The fix is either to copy the two bars and reset the note offsets or to
have a separate note accumulator for Intro sections which resets the
offsets before accumulating.