Handle rests differently
Now emit a rest as a note.  This has a pitch (nite id) of 0 and should
act as a no-op in the soundfont player.  However it acts as a
placeholder in the generated stream of notes which should allow phrasing
to act correctly, because it needs a reified object to phrase when the
rest occurs at the start of a phrase.

Changed the tests accordingly.  We now need to listen to examples.