Bar phrase break immediately after a chord symbol
This allows the tune to play in line with the accompaniment at small
phrase sizes.

The code is getting as complex as I need it now.  We have to track
through the state both whether the last note was tied and also whether
the last item encountered was a chord symbol by setting flags and
remembering to reset them.  This is incredibly fragile.